requirements.txt
langchain
torch
transformers
sentence-transformers
datasets
faiss-cpu
faiss-gpu
langchain-community
aiogram
accelerate
asyncio
nest_asyncio
python-telegram-bot